Added both buttons
Added both test buttons, the red and the green. The green button is set up to hold down the mouse click once it is pressed, and to release the mouse click once it is pressed again. The red button is set up to press the up arrow on the keyboard, which in the Unity script, will clear the water that has accumulated in the map.